some touchups + rendition #1 for downed display

+++ src/main/java/com/creatorsplash/oxygenheist/domain/player/PlayerMatchState.java
@@ -58,7 +58,7 @@ public class PlayerMatchState {
      *
      * @param amount the amount of oxygen to restore
      */
-    public void restoreOxygen(int amount) {
+    public void restoreOxygen(double amount) {
         this.oxygen = Math.min(this.maxOxygen, this.oxygen + amount);
     }
